feat(generalsio): read game updates
The game state saved in `map` and `cities` is initialized to an empty
array. I am thinking that the patch method would at the first iteration
push new items to an empty array. Although I am not sure whether it is
the best approach as patch method should not have an if checking if it
is the first time the patch function is called. Maybe we could create
init function that would be called on the first update only?

Ideally we would initialize the game states to appropriate size,
unfortunatelly, sizes are propagated with the first `game_update`. If we
used Python list then we don't have this problem as Python slices would
handle that for us. Could numpy slice work as well?