ci: Improve workflows
* fix: Align priority moves with generalsio

* fix: Make online observations include priority

* ci: Add Code Embedder by @kvankova to keep READMEs up-to-date

* refactor(exceptions): make exceptions extensible

Next step should be to utilize exceptions instead of `assert`. The main
advantage of it is that you can catch specific exception instead of
`AssertionError`.

* ci: fix code-embedder action

The manual setting of referrence breaks the checkout action. I am not
entirely sure why it was needed, or what was the intended behaviour, but
checkout action by default switches to feature branch (branch which is
used as source for PR).

* ci: rework GitHub actions

It is more of an update and simplification of triggeres and update of
some action versions.

* chore: Fix pytests

---------

Co-authored-by: Juraj Paluba <juraj.palubajp@gmail.com>